fix(prover): use u16 for ports, align include_workdir default with runtime
Port fields changed from u32 to u16 across all prover types (policy,
model, finding, queries). Prevents the prover from silently accepting
port values >65535 that the runtime rejects, which would produce
misleading PASS results on invalid policies.

Change include_workdir serde default from true to false to match the
runtime (openshell-policy uses #[serde(default)] which gives false).
The previous mismatch caused the prover to model a /sandbox path that
does not exist at runtime, producing false analysis.

Made-with: Cursor

+++ crates/openshell-prover/src/finding.rs
@@ -26,7 +26,7 @@ impl fmt::Display for RiskLevel {
 pub struct ExfilPath {
     pub binary: String,
     pub endpoint_host: String,
-    pub endpoint_port: u32,
+    pub endpoint_port: u16,
     pub mechanism: String,
     pub policy_name: String,
     /// One of `"l4_only"`, `"l7_allows_write"`, `"l7_bypassed"`.
@@ -38,7 +38,7 @@ pub struct ExfilPath {
 pub struct WriteBypassPath {
     pub binary: String,
     pub endpoint_host: String,
-    pub endpoint_port: u32,
+    pub endpoint_port: u16,
     pub policy_name: String,
     pub policy_intent: String,
     /// One of `"l4_only"`, `"l7_bypass_protocol"`, `"credential_write_scope"`.
